requirements.txt
Flask>=3.0,<4
pandas>=2.2,<3
numpy>=1.26,<3
reportlab>=4.0,<5
python-dateutil>=2.8,<3
openpyxl>=3.1,<4
pywebview>=4.4,<5
waitress>=2.1,<3
